Подключит комп-ты mainScreen и genreList к глобальному хранилищу
