+++ project/src/components/genre-list/genre-list.tsx
@@ -1,31 +1,27 @@
-import { useState } from 'react';
 import { Link } from 'react-router-dom';
 import { AppRoute } from '../../const';
-import { FilmInfo, Genre } from '../../types/types';
-import { getCatalogGenre } from '../../utils';
+import { useAppDispatch } from '../../hooks';
+import { changeGenre } from '../../store/action';
+import { FilmInfo} from '../../types/types';
+import { getGenreList } from '../../utils';
 
 type GenreListProps = {
     films: FilmInfo[];
+    activeGenre: string;
 }
 
-function GenreList({films}: GenreListProps): JSX.Element {
-  const [userGenre, setUserGenre] = useState('All genres');
+function GenreList({films, activeGenre}: GenreListProps): JSX.Element {
+  const dispatch = useAppDispatch();
 
-  const collectedGenres: string[] = [];
-  films.forEach((film) => collectedGenres.push(film.genre));
-  const uniqueGenres = ['All genres',...new Set(collectedGenres)];
-  const catalogGenres: Genre[] = [];
-  uniqueGenres.map((genre, index) =>
-    catalogGenres.push({id: index, name: getCatalogGenre(genre)},
-    ));
+  const catalogGenres = getGenreList(films);
 
   return (
     <ul className="catalog__genres-list">
       {
         catalogGenres.map((catalogGenre) =>(
-          <li key={catalogGenre.name} className={`catalog__genres-item ${catalogGenre.name === userGenre ? 'catalog__genres-item--active': ''}`}>
+          <li key={catalogGenre.name} className={`catalog__genres-item ${catalogGenre.name === activeGenre ? 'catalog__genres-item--active': ''}`}>
             <Link to={AppRoute.Main} className="catalog__genres-link"
-              onClick={() => setUserGenre(catalogGenre.name)}
+              onClick={() => dispatch(changeGenre(catalogGenre.name))}
             >{catalogGenre.name}
             </Link>
           </li>),
